+++ src/Sulu/Bundle/SecurityBundle/Admin/SecurityAdmin.php
@@ -29,7 +29,7 @@ class SecurityAdmin extends Admin
         $this->securityChecker = $securityChecker;
 
         $rootNavigationItem = new NavigationItem($title);
-        $section = new NavigationItem('');
+        $section = new NavigationItem('navigation.modules');
         $section->setPosition(20);
 
         $settings = new NavigationItem('navigation.settings');