+++ src/Sulu/Bundle/TagBundle/Admin/TagAdmin.php
@@ -29,7 +29,7 @@ class TagAdmin extends Admin
         $this->securityChecker = $securityChecker;
 
         $rootNavigationItem = new NavigationItem($title);
-        $section = new NavigationItem('');
+        $section = new NavigationItem('navigation.modules');
         $section->setPosition(20);
 
         $settings = new NavigationItem('navigation.settings');