+++ src/Sulu/Bundle/TranslateBundle/Admin/TranslateAdmin.php
@@ -21,7 +21,7 @@ class TranslateAdmin extends Admin
     {
         $rootNavigationItem = new NavigationItem($title);
 
-        $section = new NavigationItem('');
+        $section = new NavigationItem('navigation.modules');
         $section->setPosition(20);
 
         $settings = new NavigationItem('navigation.settings');